:truck: [#3825] Move categories endpoint to public folder
